Doc generation changes
