+++ src/main/java/uk/gov/hmcts/reform/adoption/adoptioncase/caseworker/event/CaseWorkerManageHearing.java
@@ -1,5 +1,6 @@
 package uk.gov.hmcts.reform.adoption.adoptioncase.caseworker.event;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -10,11 +11,25 @@ import uk.gov.hmcts.ccd.sdk.api.callback.AboutToStartOrSubmitResponse;
 import uk.gov.hmcts.ccd.sdk.type.YesOrNo;
 import uk.gov.hmcts.reform.adoption.adoptioncase.caseworker.event.page.ManageHearings;
 import uk.gov.hmcts.reform.adoption.adoptioncase.model.CaseData;
+import uk.gov.hmcts.reform.adoption.adoptioncase.model.LanguagePreference;
 import uk.gov.hmcts.reform.adoption.adoptioncase.model.ManageHearingOptions;
 import uk.gov.hmcts.reform.adoption.adoptioncase.model.State;
 import uk.gov.hmcts.reform.adoption.adoptioncase.model.UserRole;
 import uk.gov.hmcts.reform.adoption.adoptioncase.model.access.Permissions;
+import uk.gov.hmcts.reform.adoption.adoptioncase.validation.RecipientValidationUtil;
+import uk.gov.hmcts.reform.adoption.common.ccd.CcdPageConfiguration;
 import uk.gov.hmcts.reform.adoption.common.ccd.PageBuilder;
+import uk.gov.hmcts.reform.adoption.document.CaseDataDocumentService;
+
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+
+import static org.apache.commons.lang3.ObjectUtils.isEmpty;
+import static uk.gov.hmcts.reform.adoption.document.DocumentConstants.MANAGE_HEARING_NOTICES_A90;
+import static uk.gov.hmcts.reform.adoption.document.DocumentConstants.MANAGE_HEARING_NOTICES_A90_FILE_NAME;
+import static uk.gov.hmcts.reform.adoption.document.DocumentUtil.formatDocumentName;
 
 @Component
 @Slf4j
@@ -30,9 +45,13 @@ public class CaseWorkerManageHearing implements CCDConfig<CaseData, State, UserR
      */
     public static final String MANAGE_HEARINGS = "Manage hearings";
 
-    @Autowired
-    private ManageHearings manageHearings;
+    private final CcdPageConfiguration manageHearings = new ManageHearings();
 
+    @Autowired
+    ObjectMapper objectMapper;
+
+    @Autowired
+    CaseDataDocumentService caseDataDocumentService;
 
     @Override
     public void configure(ConfigBuilder<CaseData, State, UserRole> configBuilder) {
@@ -40,6 +59,15 @@ public class CaseWorkerManageHearing implements CCDConfig<CaseData, State, UserR
         log.info("Inside configure method for Event {}", CASEWORKER_MANAGE_HEARING);
         var pageBuilder = addEventConfig(configBuilder);
         manageHearings.addTo(pageBuilder);
+        pageBuilder.page("manageHearing6",this::midEventAfterRecipientSelection)
+            .showCondition("manageHearingOptions=\"addNewHearing\" OR isTheHearingNeedsRelisting=\"Yes\"")
+            .mandatory(CaseData::getRecipientsInTheCase)
+            .page("manageHearing7")
+            .showCondition("recipientsInTheCase=\"applicant1\" OR recipientsInTheCase=\"applicant2\"")
+            .label("manageHearing71","Preview Draft",null, true)
+            .optional(CaseData::getHearingA90Document)
+            .done()
+            .build();
 
     }
 
@@ -86,5 +114,45 @@ public class CaseWorkerManageHearing implements CCDConfig<CaseData, State, UserR
             .build();
     }
 
+    /*
+        This MidEvent will validate if any incorrect selection of Recipients is made.
+        In case any non-applicable Recipient is selected
+        System will throw an error.
+         */
+    public AboutToStartOrSubmitResponse<CaseData, State> midEventAfterRecipientSelection(
+        CaseDetails<CaseData, State> details,
+        CaseDetails<CaseData, State> detailsBefore
+    ) {
+        var caseData = details.getData();
+        List<String> error = new ArrayList<>();
 
+        RecipientValidationUtil.checkingApplicantRelatedSelectedRecipients(caseData, error);
+        RecipientValidationUtil.checkingChildRelatedSelectedRecipient(caseData, error);
+        RecipientValidationUtil.checkingParentRelatedSelectedRecipients(caseData, error);
+        RecipientValidationUtil.checkingOtherPersonRelatedSelectedRecipients(caseData, error);
+        RecipientValidationUtil.checkingAdoptionAgencyRelatedSelectedRecipients(caseData, error);
+
+        if (isEmpty(error)) {
+            @SuppressWarnings("unchecked")
+            Map<String, Object> templateContent = objectMapper.convertValue(caseData, Map.class);
+            log.info("templateContent {}", templateContent);
+            caseData.getManageHearingDetails().setHearingA90Document(caseDataDocumentService.renderDocument(
+                templateContent,
+                details.getId(),
+                MANAGE_HEARING_NOTICES_A90,
+                LanguagePreference.ENGLISH,
+                formatDocumentName(
+                    details.getId(),
+                    MANAGE_HEARING_NOTICES_A90_FILE_NAME,
+                    LocalDateTime.now()
+                )));
+            //TODO:  Set this value as null in aboutToSubmit callback
+            caseData.setHearingA90Document(caseData.getManageHearingDetails().getHearingA90Document());
+        }
+
+        return AboutToStartOrSubmitResponse.<CaseData, State>builder()
+            .data(caseData)
+            .errors(error)
+            .build();
+    }
 }
